Pull some things out of the demo index.js
Also, improve extracting preset names from preset urls
